refactor(pros): created _components folder and moved assigned files to it
also split the styling from old
css module to a new one
for survey page and components
